pyproject.toml
[tool.poetry]
name = "llm_library"
version = "0.1.0"
description = ""
authors = ["Gab25 <gabrielesorci.25@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
